requirements.txt
flask
flask_sqlalchemy
flask_cors
pandas
scikit-learn
lightgbm
joblib
psycopg2-binary
python-dotenv